refactor(settings): extract settings forms into settings/ directory and reorder
Split each form on the main settings page into its own file under
settings/. Payment-related forms (provider, stripe, square, webhook,
booking fee) grouped in payment.tsx since they're tightly related.

Reordered: country, email, header image, theme, public site up front;
payment group in the middle; terms, embed hosts, change password last.
