+++ src/templates/admin/settings/payment.tsx
@@ -1,0 +1,220 @@
+/**
+ * Payment Provider, Stripe, Square, and Booking Fee forms for settings
+ */
+
+import { MASK_SENTINEL } from "#lib/db/settings.ts";
+import { CsrfForm, renderFields } from "#lib/forms.tsx";
+import { Raw } from "#lib/jsx/jsx-runtime.ts";
+import type { SettingsPageState } from "#templates/admin/settings.tsx";
+import {
+  squareAccessTokenFields,
+  squareWebhookFields,
+  stripeKeyFields,
+} from "#templates/fields.ts";
+
+export const PaymentProviderForm = (s: SettingsPageState): JSX.Element => (
+  <CsrfForm
+    action="/admin/settings/payment-provider"
+    id="settings-payment-provider"
+  >
+    <h2>Payment Provider</h2>
+    <p>Choose which payment provider to use for paid events.</p>
+    <label>
+      <input
+        type="radio"
+        name="payment_provider"
+        value="none"
+        checked={!s.paymentProvider}
+      />
+      None (payments disabled)
+    </label>
+    <label>
+      <input
+        type="radio"
+        name="payment_provider"
+        value="stripe"
+        checked={s.paymentProvider === "stripe"}
+      />
+      Stripe
+    </label>
+    <label>
+      <input
+        type="radio"
+        name="payment_provider"
+        value="square"
+        checked={s.paymentProvider === "square"}
+      />
+      Square
+    </label>
+    <button type="submit">Save Payment Provider</button>
+  </CsrfForm>
+);
+
+export const StripeForm = (s: SettingsPageState): JSX.Element | null =>
+  s.paymentProvider === "stripe" ? (
+    <CsrfForm action="/admin/settings/stripe" id="settings-stripe">
+      <h2>Stripe Settings</h2>
+      <p>
+        {s.stripeKeyConfigured
+          ? "A Stripe secret key is currently configured. Enter a new key below to replace it."
+          : "No Stripe key is configured. Enter your Stripe secret key to enable Stripe payments."}
+      </p>
+      {s.stripeKeyConfigured && s.stripeKeyMode === "test" && (
+        <p class="notice warning">
+          <strong>Test mode:</strong> You are using a Stripe test key (
+          <code>sk_test_</code>). No real charges will be made. Switch to a live
+          key (<code>sk_live_</code>) when you are ready to accept real
+          payments.
+        </p>
+      )}
+      {s.stripeKeyConfigured && s.stripeKeyMode === "live" && (
+        <p class="notice">
+          <strong>Live mode:</strong> You are using a Stripe live key. Payments
+          will be charged for real.
+        </p>
+      )}
+      <p>
+        <small>
+          <a href="/admin/guide#payment-setup">Where do I find this?</a>
+        </small>
+      </p>
+      <Raw
+        html={renderFields(
+          stripeKeyFields,
+          s.stripeKeyConfigured ? { stripe_secret_key: MASK_SENTINEL } : {},
+        )}
+      />
+      <footer>
+        <button type="submit">Update Stripe Key</button>
+        {s.stripeKeyConfigured && (
+          <button type="button" id="stripe-test-btn" class="secondary">
+            Test Connection
+          </button>
+        )}
+      </footer>
+      <div id="stripe-test-result" class="hidden"></div>
+    </CsrfForm>
+  ) : null;
+
+export const SquareForm = (s: SettingsPageState): JSX.Element | null =>
+  s.paymentProvider === "square" ? (
+    <CsrfForm action="/admin/settings/square" id="settings-square">
+      <h2>Square Settings</h2>
+      <p>
+        {s.squareTokenConfigured
+          ? "A Square access token is currently configured. Enter new credentials below to replace them."
+          : "No Square access token is configured. Enter your Square credentials to enable Square payments."}
+      </p>
+      <p>
+        <small>
+          <a href="/admin/guide#payment-setup">Where do I find these?</a>
+        </small>
+      </p>
+      <Raw
+        html={renderFields(
+          squareAccessTokenFields,
+          s.squareTokenConfigured ? { square_access_token: MASK_SENTINEL } : {},
+        )}
+      />
+      <label>
+        <input
+          type="checkbox"
+          name="square_sandbox"
+          checked={s.squareSandbox}
+        />
+        Sandbox mode (use Square's test environment)
+      </label>
+      <footer>
+        <button type="submit">Update Square Credentials</button>
+        {s.squareTokenConfigured && (
+          <button type="button" id="square-test-btn" class="secondary">
+            Test Connection
+          </button>
+        )}
+      </footer>
+      <div id="square-test-result" class="hidden"></div>
+    </CsrfForm>
+  ) : null;
+
+export const SquareWebhookForm = (s: SettingsPageState): JSX.Element | null =>
+  s.paymentProvider === "square" && s.squareTokenConfigured ? (
+    <CsrfForm
+      action="/admin/settings/square-webhook"
+      id="settings-square-webhook"
+    >
+      <h2>Square Webhook</h2>
+      <p>
+        <a href="/admin/guide#payment-setup">See the full setup guide</a>
+      </p>
+      <article>
+        <aside>
+          <p>
+            To receive payment notifications, set up a webhook in your Square
+            Developer Dashboard:
+          </p>
+          <ol>
+            <li>
+              Go to your <strong>Square Developer Dashboard</strong> and select
+              your application
+            </li>
+            <li>
+              Navigate to <strong>Webhooks</strong> in the left sidebar
+            </li>
+            <li>
+              Click <strong>Add Subscription</strong>
+            </li>
+            <li>
+              Set the <strong>Notification URL</strong> to:
+              <br />
+              <code>{s.webhookUrl}</code>
+            </li>
+            <li>
+              Subscribe to the <strong>payment.updated</strong> event
+            </li>
+            <li>
+              Save the subscription and copy the <strong>Signature Key</strong>
+            </li>
+            <li>Paste the signature key below</li>
+          </ol>
+        </aside>
+      </article>
+      <p>
+        {s.squareWebhookConfigured
+          ? "A webhook signature key is currently configured. Enter a new key below to replace it."
+          : "No webhook signature key is configured. Follow the steps above to set one up."}
+      </p>
+      <Raw
+        html={renderFields(
+          squareWebhookFields,
+          s.squareWebhookConfigured
+            ? { square_webhook_signature_key: MASK_SENTINEL }
+            : {},
+        )}
+      />
+      <button type="submit">Update Webhook Key</button>
+    </CsrfForm>
+  ) : null;
+
+export const BookingFeeForm = (s: SettingsPageState): JSX.Element | null =>
+  s.paymentProvider ? (
+    <CsrfForm action="/admin/settings/booking-fee" id="settings-booking-fee">
+      <h2>Booking Fee</h2>
+      <p>
+        Percentage fee added at checkout (e.g. 1.5 for 1.5%). Set to 0 to
+        disable. Max 10.
+      </p>
+      <label>
+        Booking Fee (%)
+        <input
+          type="number"
+          name="booking_fee"
+          step="0.1"
+          min="0"
+          max="10"
+          value={s.bookingFee}
+          required
+        />
+      </label>
+      <button type="submit">Save Booking Fee</button>
+    </CsrfForm>
+  ) : null;
